Handle fragment and key error events in base-stream controller
